Extract readLicenseViaALPR platerecognizer functionality to `alpr.js`
