+++ src/alpr.js
@@ -1,0 +1,111 @@
+import exifr from 'exifr/dist/full.umd.js';
+import sharp from 'sharp';
+import nodeFetch from 'node-fetch';
+import FormData from 'form-data';
+
+async function orientImageBuffer({ attachmentBuffer }) {
+  console.info(
+    'Extracted GPS latitude/longitude location from EXIF metadata',
+    await exifr.gps(attachmentBuffer),
+  ); // eslint-disable-line no-console
+
+  // eslint-disable-next-line no-console
+  console.log(
+    await exifr.parse(attachmentBuffer, ['CreateDate', 'OffsetTimeDigitized']),
+  );
+
+  console.time(`orientImageBuffer`); // eslint-disable-line no-console
+  // eslint-disable-next-line no-console
+  console.log(
+    `image buffer length BEFORE sharp: ${attachmentBuffer.length} bytes`,
+  );
+  return sharp(attachmentBuffer)
+    .rotate()
+    .toBuffer()
+    .catch(() => attachmentBuffer)
+    .then(buffer => Buffer.from(buffer))
+    .then(async buffer => {
+      console.info(
+        'Extracted GPS latitude/longitude location from EXIF metadata',
+        await exifr.gps(buffer),
+      ); // eslint-disable-line no-console
+
+      // eslint-disable-next-line no-console
+      console.log(
+        await exifr.parse(buffer, ['CreateDate', 'OffsetTimeDigitized']),
+      );
+
+      console.log(`image buffer length AFTER sharp: ${buffer.length} bytes`); // eslint-disable-line no-console
+      console.timeEnd(`orientImageBuffer`); // eslint-disable-line no-console
+      return buffer;
+    });
+}
+
+// https://app.platerecognizer.com/upload-limit/
+const downscaleForPlateRecognizer = buffer => {
+  const fileSize = buffer.length;
+  const maxFilesize = 2411654;
+
+  if (fileSize >= maxFilesize) {
+    const targetWidth = 4096;
+
+    // eslint-disable-next-line no-console
+    console.log(
+      `file size is greater than maximum of ${maxFilesize} bytes, attempting to scale down to width of ${targetWidth}`,
+    );
+
+    return sharp(buffer)
+      .resize({ width: targetWidth })
+      .toBuffer()
+      .catch(error => {
+        console.error('could not scale down, using unscaled image', { error });
+        return buffer;
+      })
+      .then(resizedBufferish => {
+        const resizedBuffer = Buffer.from(resizedBufferish);
+        // eslint-disable-next-line no-console
+        console.log(
+          `file size after scaling down: ${resizedBuffer.length} bytes`,
+        );
+        return resizedBuffer;
+      });
+  }
+
+  return buffer;
+};
+
+export default function readLicenseViaALPR({
+  attachmentBuffer,
+  PLATERECOGNIZER_TOKEN,
+}) {
+  return orientImageBuffer({ attachmentBuffer })
+    .then(downscaleForPlateRecognizer)
+    .then(buffer => buffer.toString('base64'))
+    .then(attachmentBytesRotated => {
+      console.log('STARTING platerecognizer'); // eslint-disable-line no-console
+      console.time(`/platerecognizer plate-reader`); // eslint-disable-line no-console
+
+      const body = new FormData();
+
+      body.append('upload', attachmentBytesRotated);
+
+      // body.append("regions", "us-ny"); // Change to your country
+      body.append('regions', 'us'); // Change to your country
+
+      return nodeFetch('https://api.platerecognizer.com/v1/plate-reader/', {
+        method: 'POST',
+        headers: {
+          Authorization: `Token ${PLATERECOGNIZER_TOKEN}`,
+        },
+        body,
+      })
+        .then(platerecognizerRes => {
+          console.info('/platerecognizer plate-reader', {
+            platerecognizerRes,
+          });
+          return platerecognizerRes;
+        })
+        .then(platerecognizerRes => platerecognizerRes.json())
+        .finally(() => console.timeEnd(`/platerecognizer plate-reader`)); // eslint-disable-line no-console
+    });
+}
